fixes errors, little changes

+++ src/components/projectile_component/mod.rs
@@ -11,7 +11,7 @@ pub struct Projectile {
     pub passive_skills: Vec<Skill>,
 
     pub starting_position: Position<i32>,
-    pub motion_coefficient: Position<i32>,
+    pub motion_coefficient: Position<f32>,
     pub range: u8,
     pub is_missed: bool,
     pub can_pierce: bool,
